backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# ML/CV
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.0.200
boxmot>=10.0.47
opencv-python>=4.10.0
numpy>=1.24.3
pillow>=10.1.0

# Storage & Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
python-jose[cryptography]==3.3.0
aiofiles==23.2.1

# Authentication
passlib[bcrypt]==1.7.4

# Real-Time Streaming & WebSocket
websockets>=12.0

# Face Recognition
deepface>=0.0.79
tf-keras>=2.16.0

# Audio Processing
openai-whisper>=20231117
spacy>=3.7.0

# Gesture Recognition
mediapipe>=0.10.0
dtaidistance>=2.3.10

# LLM Integration
anthropic>=0.18.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
